refactor(bin/publish-npm-modules.sh): support 'npm link' with '--link'

+++ bin/publish-npm-modules.sh
@@ -8,7 +8,8 @@ declare args=()
 declare dry_run=0
 declare only_platforms=0
 declare only_top_level=0
-
+declare remove_socket_home=1
+declare do_global_link=0
 
 if [[ "$platform" = "linux" ]]; then
   if [ -n "$WSL_DISTRO_NAME" ] || uname -r | grep 'Microsoft'; then
@@ -40,14 +41,32 @@ while (( $# > 0 )); do
     continue
   fi
 
+  if [[ "$arg" = "--no-remove-socket-home" ]]; then
+    remove_socket_home=0
+    continue
+  fi
+
+  if [[ "$arg" = "--link" ]]; then
+    do_global_link=1
+    continue
+  fi
+
   args+=("$arg")
 done
 
-rm -rf "$SOCKET_HOME"
+if (( remove_socket_home )); then
+  rm -rf "$SOCKET_HOME"
+fi
+
 mkdir -p "$SOCKET_HOME"
 
 export SOCKET_HOME
 export PREFIX
+
+if (( do_global_link && !dry_run )); then
+  echo >&2 "warning: '--link' implies '--dry-run' too"
+  dry_run=1
+fi
 
 if (( !only_top_level )); then
   "$root/bin/install.sh" || exit $?
@@ -111,6 +130,10 @@ if (( !only_top_level )); then
       # echo "# npm publish ${args[@]}"
       npm pack "${args[@]}" || exit $?
     fi
+
+    if (( do_global_link )); then
+      npm link
+    fi
   done
 fi
 
@@ -123,5 +146,14 @@ if (( !only_platforms || only_top_level )); then
   else
     # echo "# npm publish ${args[@]}"
     npm pack "${args[@]}" || exit $?
+  fi
+
+  if (( do_global_link )); then
+    for arch in "${archs[@]}"; do
+      declare package="@socketsupply/socket-$platform-${arch/x86_64/x64}"
+      npm link "$package"
+    done
+
+    npm link
   fi
 fi
